Rename types navigation package to use dotcom- instead of anvil-
