+++ packages/anvil-ui-ft-footer/src/__stories__/footer.ts
@@ -1,4 +1,4 @@
-import { TNavMenu } from '../../../anvil-types-navigation'
+import { TNavMenu } from '../../../dotcom-types-navigation'
 
 const data: TNavMenu = {
   label: 'Footer',
