+++ packages/anvil-ui-ft-footer/src/components/partials.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { TNavMenuItem } from '@financial-times/anvil-types-navigation'
+import { TNavMenuItem } from '@financial-times/dotcom-types-navigation'
 
 type TLinkProps = {
   item: TNavMenuItem
