+++ packages/anvil-ui-ft-footer/src/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { TNavigationData } from '@financial-times/anvil-types-navigation'
+import { TNavigationData } from '@financial-times/dotcom-types-navigation'
 import {
   FooterContents,
   MoreFromFT,
