+++ packages/anvil-ui-ft-layout/package.json
@@ -24,7 +24,7 @@
   "dependencies": {
     "@financial-times/anvil-ui-ft-header": "file:../anvil-ui-ft-header",
     "@financial-times/anvil-ui-ft-footer": "file:../anvil-ui-ft-footer",
-    "@financial-times/anvil-types-navigation": "file:../anvil-types-navigation",
+    "@financial-times/dotcom-types-navigation": "file:../dotcom-types-navigation",
     "focus-visible": "^4.1.5"
   },
   "peerDependencies": {
