+++ packages/anvil-ui-ft-layout/src/components/Layout.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Header, LogoOnly, Drawer, THeaderOptions } from '@financial-times/anvil-ui-ft-header/component'
-import { TNavigationData } from '@financial-times/anvil-types-navigation'
+import { TNavigationData } from '@financial-times/dotcom-types-navigation'
 import { Footer, LegalFooter, TFooterOptions } from '@financial-times/anvil-ui-ft-footer/component'
 import { loadCustomFontsJS } from '../font-loading'
 import Template from './Template'
