+++ packages/dotcom-middleware-navigation/src/navigation.ts
@@ -1,5 +1,5 @@
 import { Request, Response, NextFunction } from 'express'
-import { TNavigationData } from '@financial-times/anvil-types-navigation'
+import { TNavigationData } from '@financial-times/dotcom-types-navigation'
 import { Navigation, TNavOptions } from '@financial-times/dotcom-server-navigation'
 import handleEdition from './handleEdition'
