+++ packages/dotcom-types-navigation/.npmignore
