+++ packages/dotcom-types-navigation/README.md
@@ -1,9 +1,9 @@
-# @financial-times/anvil-types-navigation
+# @financial-times/dotcom-types-navigation
 
 Generic type definitions for packages which consume navigation data from the [Origami Navigation Service](https://www.ft.com/__origami/service/navigation/v2).
 
 ## Installation
 
 ```sh
-npm install -S @financial-times/anvil-types-navigation
+npm install -S @financial-times/dotcom-types-navigation
 ```
